+++ ext/standard/quot_print.c
@@ -53,7 +53,7 @@ static char php_hex2int(int c) /* {{{ */
 
 PHPAPI zend_string *php_quot_print_decode(const unsigned char *str, size_t length, int replace_us_by_ws) /* {{{ */
 {
-	register unsigned int i;
+	register size_t i;
 	register unsigned const char *p1;
 	register unsigned char *p2;
 	register unsigned int h_nbl, l_nbl;
